Match d_a_obj_ice_s rideCallBack (#2977)
* Match d_a_obj_ice_s rideCallBack


* Fix debug regression

---------

+++ src/PowerPC_EABI_Support/MSL/MSL_C/MSL_Common/Src/float.c
@@ -1,13 +1,13 @@
 #include "float.h"
 
-unsigned long __float_nan[] = {0x7FFFFFFF};
+int __float_nan[] = {0x7FFFFFFF};
 
-unsigned long __float_huge[] = {0x7F800000};
+int __float_huge[] = {0x7F800000};
 
 #if !__REVOLUTION_SDK__
-unsigned long __float_max[] = {0x7F7FFFFF};
+int __float_max[] = {0x7F7FFFFF};
 
-unsigned long __float_epsilon[] = {0x34000000};
+int __float_epsilon[] = {0x34000000};
 #endif
 
 #if DEBUG
